% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\name{as}
\alias{as}
\title{As("ChipInfo", "GRanges")}
\description{
As("ChipInfo", "GRanges")

As("ChipInfo", "GRanges")

As("ChipInfo", "GRanges")

As("GRanges", "ChipInfo")

As("RangedData", "ChipInfo")

As("data.frame", "ChipInfo")

As("data.frame", "RangedData")

As("data.frame", "GRanges")

Note that for automatic conversion of a data.frame to RangedData/GRanges, a column named 'chr' 
or 'seqnames' in the data.frame is expected/required to make the conversion effectively. 
Otherwise use 'ranged.to.data.frame()'

As("GRanges", "data.frame")
}

