% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{chip}
\alias{chip}
\alias{chip,ChipInfo-method}
\title{Retrieve the Chip name for ChipInfo}
\usage{
chip(x)

\S4method{chip}{ChipInfo}(x)
}
\arguments{
\item{x}{a ChipInfo object}
}
\value{
character string
}
\description{
Simply returns the name of the chip, e.g, 'ImmunoChip'
}
