% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{chipId}
\alias{chipId}
\alias{chipId,ChipInfo-method}
\title{Access chip-ids for ChipInfo}
\usage{
chipId(x)

\S4method{chipId}{ChipInfo}(x)
}
\arguments{
\item{x}{a ChipInfo object}
}
\value{
chip ids: character vector of IDs (or NAs)
}
\description{
Returns the chip-ids for the chip object, e.g, "imm_1_898835", etc
Only if these are annotated internally, or else a vector of NAs
Note that the main purpose of this is because sometimes chip-ids
do not satisfy conditions to be an R column/row name, e.g, start
with a number, illegal characters, etc. So this allows certain
functions to return the actual chip names that would match the 
official manifest. These will largely be the same as the rownames,
but the rownames will always be valid R column names, converted from
the original using clean.snp.ids() [internal function]
}

