% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{[[,ChipInfo,ANY,ANY-method}
\alias{[[,ChipInfo,ANY,ANY-method}
\title{Subset ChipInfo by chromosome}
\usage{
\S4method{[[}{ChipInfo,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{a ChipInfo object}

\item{i}{a chromosome number or letter, i.e, one of seqlevels(x)}

\item{j}{always leave missing, not applicable for this method.}

\item{...}{further arguments - again there should not be any}
}
\value{
ChipInfo object for the subset of SNPs on chromosome i
}
\description{
Returns the subset of the ChipInfo object for which SNPs are on
the chromosome specified, by either number or character.
}

