% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{chrIndices,RangedData-method}
\alias{chrIndices,RangedData-method}
\title{Chromosome indices method for ranged objects}
\usage{
\S4method{chrIndices}{RangedData}(object)
}
\arguments{
\item{object}{RangedData or GRanges object}
}
\value{
matrix of indexes, colnames  first, last, offset
}
\description{
Return the list of chromosome start and end indexes from a RangedData object
}
