% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{chrm}
\alias{chrm}
\alias{chrm,ChipInfo-method}
\alias{chrm,GRanges-method}
\alias{chrm,RangedData-method}
\title{Chromosome method for RangedData objects}
\usage{
chrm(object)

\S4method{chrm}{RangedData}(object)

\S4method{chrm}{GRanges}(object)

\S4method{chrm}{ChipInfo}(object)
}
\arguments{
\item{object}{RangedData object}
}
\value{
vector of chromosome values for each range/SNP
}
\description{
Return the list of chromosome values from a RangedData object
}

