% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{chrNames,RangedData-method}
\alias{chrNames,RangedData-method}
\title{Chromosome names method for ranged objects}
\usage{
\S4method{chrNames}{RangedData}(object)
}
\arguments{
\item{object}{RangedData or GRanges object}
}
\value{
vector of names
}
\description{
Return the list of chromosome labels from a RangedData object
}
