% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{chrSel}
\alias{chrSel}
\alias{chrSel,RangedData-method}
\alias{chrSel,GRanges-method}
\alias{chrSel,ChipInfo-method}
\title{Select chromosome subset for ranged objects}
\usage{
chrSel(object, chr)

\S4method{chrSel}{RangedData}(object, chr)

\S4method{chrSel}{GRanges}(object, chr)

\S4method{chrSel}{ChipInfo}(object, chr)
}
\arguments{
\item{object}{a ChipInfo, GRanges or RangedData object}

\item{chr}{vector, string or numeric of which chromosome(s) to select}
}
\value{
vector of chromosome values for each range/SNP
}
\description{
Returns the object filtered for specific chromosomes for a ranged object

Returns the object filtered for specific chromosomes for a RangedData object

Returns the object filtered for specific chromosomes for a GRanges object

Returns the object filtered for specific chromosomes for a GRanges object
}
