% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{convTo37}
\alias{convTo36}
\alias{convTo36,ChipInfo-method}
\alias{convTo37}
\alias{convTo37,ChipInfo-method}
\alias{convTo38}
\alias{convTo38,ChipInfo-method}
\title{Convert ChipInfo between build 36/37/38 coordinates}
\usage{
convTo37(x)

\S4method{convTo37}{ChipInfo}(x)

convTo36(x)

\S4method{convTo36}{ChipInfo}(x)

convTo38(x)

\S4method{convTo38}{ChipInfo}(x)
}
\arguments{
\item{x}{a ChipInfo object}
}
\value{
convTo37: Returns a ChipInfo object with the build updated to hg19 coordinates

convTo36: Returns a ChipInfo object with the build updated to hg18 coordinates

convTo38: Returns a ChipInfo object with the build updated to hg38 coordinates
}
\description{
Returns the a ChipInfo object with positions updated to build 36/37/38
coordinates, assuming that the build() slot was entered correctly. 
Ensure that the value of ucsc(x) is correct before
running this function for conversion; for instance, if the coordinates 
are already build 37/hg19, but ucsc(x)!="hg19" (incorrect value), then
these coordinates will be transformed in a relative manner rendering the
result meaningless.
}

