% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{extraColumnSlotNames2,ANY-method}
\alias{extraColumnSlotNames2,ANY-method}
\title{Get extra column slot names}
\usage{
\S4method{extraColumnSlotNames2}{ANY}(x)
}
\arguments{
\item{x}{any object}
}
\value{
internal function
}
\description{
Return the list of chromosome start and end indexes from a RangedData object
}
