% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{rownames,GRanges-method}
\alias{rownames,GRanges-method}
\alias{rownames<-,GRanges-method}
\title{rownames method for GRanges objects}
\usage{
\S4method{rownames}{GRanges}(x)

\S4method{rownames}{GRanges}(x) <- value
}
\arguments{
\item{x}{a GRanges object}

\item{value}{a character string}
}
\value{
rownames: Character vector of row names (e.g, SNP IDs).
}
\description{
rownames: Returns the row names.
}
