\name{humarray-internal}

\alias{do.cw}
\alias{.rownames3}
\alias{makePrettyMatrixForCompactPrinting2}
\alias{.makeNakedMatFromChipInfo}
\alias{extraColumnSlotNames2}
\alias{extraColumnSlots2}
\alias{finitize}
\alias{minna}
\alias{maxna}
\alias{meanna}
\alias{medianna}
\alias{sdna}
\alias{sumna}
\alias{sortna}
\alias{l10}
\alias{Cor}
\alias{pt2}
\alias{ucsc.sanitizer}
\alias{pduplicated}
\alias{comma}
\alias{order.chr}
\alias{sort.chr}
\alias{clean.snp.ids}
\alias{hard.coded.conv}
\alias{rmv.trail}
\alias{plot.get.area}
\alias{add.trail}
\alias{rsampid}
\alias{rsnpid}
\alias{emd.rmv}
\alias{chrOrder2}
\alias{chrNames2}
\alias{toGenomeOrder2}
\alias{chrInfo2}
\alias{chrIndices2}
\alias{chr2}
\alias{make.divisor}
\alias{plotdf}
\alias{chrnums.to.txt}
\alias{chrnames.to.num}
\alias{tidy.extra.chr}
\alias{gene.duplicate.report}
\alias{validate.dir.for}
\alias{substitute.36s}
\alias{TGORD}
\alias{TGOGR}
\alias{egSymb}
\alias{hg18ToHg19}
\alias{hg19ToHg18}
\alias{hg19ToHg38}
\alias{hg38ToHg19}
\alias{.updateAllele}

\title{Internal humarray Functions}
\description{
  Internal humarray functions
}
\details{
  These are not recommended/needed to be called by the user 
}
\keyword{ internal }
