% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{plot,GRanges,ANY-method}
\alias{plot,GRanges,ANY-method}
\alias{GRanges}
\alias{GRanges-method}
\alias{plot,RangedData,ANY-method}
\alias{RangedData}
\alias{RangedData-method}
\title{Plot method for GRanges objects}
\usage{
\S4method{plot}{GRanges,ANY}(x, y, ...)

\S4method{plot}{RangedData,ANY}(x, y, ...)
}
\arguments{
\item{x}{a GRanges or RangedData object}

\item{y}{not used for plotRanges}

\item{...}{further arguments, see plotRanges()}
}
\description{
See plotRanges()

Plot method for RangedData objects
}
\seealso{
\code{\link{plotRanges}}
}
