% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humarray.R
\docType{methods}
\name{rs.id}
\alias{rs.id}
\alias{rs.id,ChipInfo-method}
\title{Access rs-ids for ChipInfo}
\usage{
rs.id(x, b = TRUE)

\S4method{rs.id}{ChipInfo}(x, b = TRUE)
}
\arguments{
\item{x}{a ChipInfo object}

\item{b}{logical, whether to show 'b' suffixes on rs.ids which
are created in the background to allow duplicate ids to be uniquely
represented for lookup and reference purposes.}
}
\value{
rs-ids: character vector of IDs (or NAs)
}
\description{
Returns the rs-ids for the chip object, e.g, "rs689", etc
Only if these are annotated internally, or else a vector of NAs
}

